+++ src/main/java/com/github/mjkuranda/spaceadventure2/entities/Entity.java
@@ -34,7 +34,15 @@ public abstract class Entity implements Moveable, Destroyable, Damageable {
 
     @Override
     public void move() {
-        y += speed;
+        if (turn == EntityTurn.INCOMING) {
+            y += speed;
+
+            return;
+        }
+
+        if (turn == EntityTurn.OUTCOMING) {
+            y -= speed;
+        }
     }
 
     public void move(EntityDirection direction) {
@@ -63,6 +71,17 @@ public abstract class Entity implements Moveable, Destroyable, Damageable {
         }
     }
 
+    public boolean collides(Entity e) {
+        if (e == null) {
+            return false;
+        }
+
+        float diffX = Math.abs(this.getX() - e.getX());
+        float diffY = Math.abs(this.getY() - e.getY());
+
+        return diffX > 0 && diffX < 1 && diffY > 0 && diffY < 1;
+    }
+
     public void setCoords(float x, float y) {
         this.x = x;
         this.y = y;
@@ -78,6 +97,10 @@ public abstract class Entity implements Moveable, Destroyable, Damageable {
 
     public void setSubscriber(LinkedList<Entity> subscriber) {
         this.subscriber = subscriber;
+    }
+
+    public void setTurn(EntityTurn turn) {
+        this.turn = turn;
     }
 
     public LinkedList<Entity> getSubscriber() {