+++ src/main/java/com/github/mjkuranda/spaceadventure2/entities/SpaceShip.java
@@ -1,6 +1,5 @@
 package com.github.mjkuranda.spaceadventure2.entities;
 
-import com.github.mjkuranda.spaceadventure2.entities.missiles.Missile;
 import com.github.mjkuranda.spaceadventure2.entities.missiles.MissileType;
 import com.github.mjkuranda.spaceadventure2.entities.shootable.Shooting;
 
@@ -8,7 +7,7 @@ import java.util.LinkedList;
 
 public class SpaceShip extends SpaceEntity {
 
-    public SpaceShip(LinkedList<Missile> missileSubscriber, float x, float y) {
+    public SpaceShip(LinkedList<Entity> missileSubscriber, float x, float y) {
         super(EntityType.SPACESHIP, null, EntityTurn.OUTCOMING, new Shooting(missileSubscriber, MissileType.LASER), x, y, 0.2f, 100);
     }
 }