+++ src/main/java/com/github/mjkuranda/spaceadventure2/entities/missiles/LaserMissile.java
@@ -9,6 +9,10 @@ import java.util.LinkedList;
 public class LaserMissile extends Missile {
 
     public LaserMissile(LinkedList<Entity> subscriber, EntityTurn turn, float x, float y) {
-        super(EntityType.LASER_MISSILE, subscriber, turn, x, y, 10, 100, 50);
+        super(EntityType.LASER_MISSILE, subscriber, turn, x, y, 0.3f, 100, 50);
+    }
+
+    public LaserMissile() {
+        this(null, null, -1, -1);
     }
 }