+++ src/main/java/com/github/mjkuranda/spaceadventure2/entities/shootable/Shooting.java
@@ -1,5 +1,6 @@
 package com.github.mjkuranda.spaceadventure2.entities.shootable;
 
+import com.github.mjkuranda.spaceadventure2.entities.Entity;
 import com.github.mjkuranda.spaceadventure2.entities.missiles.Missile;
 import com.github.mjkuranda.spaceadventure2.entities.missiles.MissileType;
 
@@ -7,11 +8,11 @@ import java.util.LinkedList;
 
 public class Shooting implements Shootable {
 
-    private LinkedList<Missile> subscriber;
+    private LinkedList<Entity> subscriber;
 
     private MissileType missileType;
 
-    public Shooting(LinkedList<Missile> subscriber, MissileType missileType) {
+    public Shooting(LinkedList<Entity> subscriber, MissileType missileType) {
         this.subscriber = subscriber;
         this.missileType = missileType;
     }