+++ src/main/java/com/github/mjkuranda/spaceadventure2/map/GameMap.java
@@ -1,7 +1,9 @@
 package com.github.mjkuranda.spaceadventure2.map;
 
 import com.github.mjkuranda.spaceadventure2.entities.*;
+import com.github.mjkuranda.spaceadventure2.entities.missiles.LaserMissile;
 import com.github.mjkuranda.spaceadventure2.entities.missiles.Missile;
+import com.github.mjkuranda.spaceadventure2.entities.missiles.MissileType;
 import com.github.mjkuranda.spaceadventure2.states.StatesId;
 import org.newdawn.slick.GameContainer;
 import org.newdawn.slick.Input;
@@ -20,12 +22,12 @@ public class GameMap {
 
     /** All lines of map */
     private LinkedList<Entity>[] entityLines;
-    private LinkedList<Missile> missileLines;
+    private LinkedList<Entity> playerMissiles;
 
     public GameMap() {
         initLists();
 
-        player = new SpaceShip(missileLines, X_SIZE / 2.0f, Y_SIZE - 1);
+        player = new SpaceShip(playerMissiles, X_SIZE / 2.0f, Y_SIZE - 1);
     }
 
     public void spawn(EntityType type) {
@@ -39,6 +41,15 @@ public class GameMap {
         entity.setSubscriber(entityLines[xInt]);
 
         entityLines[xInt].add(entity);
+    }
+
+    public void spawn(MissileType type) {
+        Missile missile = getMissile(type);
+        missile.setCoords(player.getX(), player.getY() - 1);
+        missile.setSubscriber(playerMissiles);
+        missile.setTurn(EntityTurn.OUTCOMING);
+
+        playerMissiles.add(missile);
     }
 
     public void update(GameContainer container, StateBasedGame game) {
@@ -58,6 +69,46 @@ public class GameMap {
             }
         }
 
+        /** Update player missiles */
+        var missileIt = playerMissiles.iterator();
+
+        while (missileIt.hasNext()) {
+            var missile = missileIt.next();
+
+            missile.move();
+
+            int x = (int) missile.getX();
+
+            if (missile.collides(getLast(x))) {
+                missileIt.remove();
+                getLast(x).destroy();
+                missile.destroy();
+
+                continue;
+            }
+
+            if (missile.collides(getLast(x - 1))) {
+                missileIt.remove();
+                getLast(x - 1).destroy();
+                missile.destroy();
+
+                continue;
+            }
+
+            if (missile.collides(getLast(x + 1))) {
+                missileIt.remove();
+                getLast(x + 1).destroy();
+                missile.destroy();
+
+                continue;
+            }
+
+            if (isOutOfMap(missile)) {
+                missileIt.remove();
+                missile.destroy();
+            }
+        }
+
         /** Player update */
         Input in = container.getInput();
 
@@ -71,6 +122,10 @@ public class GameMap {
 
         if (in.isKeyPressed(Input.KEY_G)) {
             spawn(EntityType.ASTEROID);
+        }
+
+        if (in.isKeyPressed(Input.KEY_SPACE)) {
+            spawn(MissileType.LASER);
         }
 
         if (playerCollides()) {
@@ -93,6 +148,10 @@ public class GameMap {
         return entityLines;
     }
 
+    public LinkedList<Entity> getPlayerMissiles() {
+        return playerMissiles;
+    }
+
     private Entity getEntity(EntityType type) {
         return switch(type) {
             case ASTEROID -> new Asteroid();
@@ -102,13 +161,20 @@ public class GameMap {
         };
     }
 
+    private Missile getMissile(MissileType type) {
+        return switch (type) {
+            case LASER -> new LaserMissile();
+            default -> null;
+        };
+    }
+
     private boolean isOutOfMap(Entity e) {
-        return e.getY() * 32 > Y_SIZE * 32;
+        return e.getY() * 32 > Y_SIZE * 32 || e.getY() < 0;
     }
 
     private void initLists() {
         entityLines = new LinkedList[Y_SIZE];
-        missileLines = new LinkedList<>();
+        playerMissiles = new LinkedList<>();
 
         initList(entityLines);
     }
@@ -117,6 +183,14 @@ public class GameMap {
         for (int i = 0; i < listLines.length; i++) {
             listLines[i] = new LinkedList<>();
         }
+    }
+
+    private Entity getLast(int x) {
+        if (entityLines[x].size() == 0) {
+            return null;
+        }
+
+        return entityLines[x].getLast();
     }
 
     private boolean playerCollides() {